Allow For Overtime Play in "Game Over" Function
In the original commit, the game was ended at the bottom of the ninth, even if there was a tie score. Code in the "Game Over" function that handled overtime was commented out. This revision removes the comments in favor of the overtime code. Games will now end only when a team has scored and won, regardless of innings.